pyproject.toml
[tool.poetry]
name = "newsAPI"
version = "0.1.0"
description = ""
authors = ["farahzuot <farahz997@eng.hu.edu.jo>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.1"
psycopg2-binary = "^2.9.1"
django-environ = "^0.4.5"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
